useState constants, handleSearch function
In the App.js file I imported useSearchBook custom hook and useState so I set a query and pageNum variable or state. The handleSearch function was created and in its param there is an e for event which is placed in the setQuery method or function for the useState the setPageNum function has a 1 in its params. The useSearchBook gets the query and pageNum from its param when text is entered into the input box..the event for the handleSearch is someone typing into the box then setQuery gets the data and stores it in query